Added feature to make change in value of Actuators(high/low/auto)

+++ routes/userRoutes.js
@@ -1,11 +1,10 @@
 const express = require("express");
-const { getAllUsers, signUp, deleteAll, protectRoute, getUserProfile, login } = require("../controllers/userController");
+const { getAllUsers, signUp, deleteAll, protectRoute, getUserProfile, login, isAuthorised, revokeAccess } = require("../controllers/userController");
 const router = express.Router();
 const passport = require('passport');
 const { signin, signinCallback, dashboard } = require("../controllers/googleAuthController");
 const downloadData = require("../controllers/nodemailerController");
 
-router.get('/allUsers', getAllUsers);
 router.post('/signup', signUp);
 router.post('/login', login);
 router.delete('/delete', deleteAll)
@@ -23,5 +22,8 @@ router
     .route('/userProfile')
     .get(getUserProfile)
 
-router.get('/download',downloadData);
+router.get('/download', downloadData);
+router.use(isAuthorised(['admin']));
+router.get('/allUsers', getAllUsers);
+router.post('/revokeAccess',revokeAccess)
 module.exports = router;
